requirements.txt
streamlit==1.31.0
plotly==5.18.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.2
matplotlib==3.7.2
seaborn==0.12.2
altair==5.2.0
statsmodels==0.14.0
setuptools>=68.0.0
wheel>=0.40.0
chromadb==0.4.22
google-generativeai==0.3.2
python-dotenv==1.0.0
langchain==0.1.9
pydantic==2.6.1 